Use nint in for loop

+++ src/ImageSharp/Formats/Webp/Lossy/QuantEnc.cs
@@ -726,7 +726,7 @@ namespace SixLabors.ImageSharp.Formats.Webp.Lossy
         {
             uint v = src[0] * 0x01010101u;
             Span<byte> vSpan = BitConverter.GetBytes(v).AsSpan();
-            for (int i = 0; i < 16; i++)
+            for (nint i = 0; i < 16; i++)
             {
                 if (!src.Slice(0, 4).SequenceEqual(vSpan) || !src.Slice(4, 4).SequenceEqual(vSpan) ||
                     !src.Slice(8, 4).SequenceEqual(vSpan) || !src.Slice(12, 4).SequenceEqual(vSpan))
@@ -748,7 +748,7 @@ namespace SixLabors.ImageSharp.Formats.Webp.Lossy
             int offset = 0;
             while (numBlocks-- > 0)
             {
-                for (int i = 1; i < 16; i++)
+                for (nint i = 1; i < 16; i++)
                 {
                     // omit DC, we're only interested in AC
                     score += Unsafe.Add(ref levelsRef, offset) != 0 ? 1 : 0;
